requirements.txt
requests
python-dotenv
azure-identity
azure-keyvault-secrets
azure-communication-email
openai